refactor: MapPlaceProjectionDto 클래스에서 레코드로 변경
