+++ src/main/java/sws/songpin/domain/place/dto/projection/MapPlaceProjectionDto.java
@@ -1,0 +1,15 @@
+package sws.songpin.domain.place.dto.projection;
+
+import sws.songpin.domain.genre.entity.GenreName;
+
+import java.time.LocalDate;
+
+public record MapPlaceProjectionDto (
+    Long placeId,
+    Double latitude,
+    double longitude,
+    Long placePinCount,
+    LocalDate listenedDate,
+    Long songId,
+    GenreName genreName
+) { }
