+++ src/main/java/sws/songpin/domain/place/dto/response/MapPlaceUnitDto.java
@@ -1,6 +1,7 @@
 package sws.songpin.domain.place.dto.response;
 
 import sws.songpin.domain.genre.entity.GenreName;
+import sws.songpin.domain.place.dto.projection.MapPlaceProjectionDto;
 
 public record MapPlaceUnitDto(
         Long placeId,
@@ -11,6 +12,6 @@ public record MapPlaceUnitDto(
         GenreName genreName
 ){
     public static MapPlaceUnitDto from(MapPlaceProjectionDto dto) {
-        return new MapPlaceUnitDto(dto.getPlaceId(), dto.getLatitude(), dto.getLongitude(), dto.getPlacePinCount(), dto.getSongId(), dto.getGenreName());
+        return new MapPlaceUnitDto(dto.placeId(), dto.latitude(), dto.longitude(), dto.placePinCount(), dto.songId(), dto.genreName());
     }
 }
